front: move AddOrEditStudyModal in applications/operationalStudies
