+++ front/src/applications/operationalStudies/_operationalStudies.scss
@@ -1,0 +1,1 @@
+@use './components/Study/addOrEditStudyModal';
