+++ front/src/applications/operationalStudies/components/Study/_addOrEditStudyModal.scss
