+++ front/src/applications/operationalStudies/views/Study/StudyView.tsx
@@ -7,6 +7,7 @@ import { useParams } from 'react-router-dom';
 import AddNewCard from 'applications/operationalStudies/components/AddNewCard';
 import BreadCrumbs from 'applications/operationalStudies/components/BreadCrumbs';
 import FilterTextField from 'applications/operationalStudies/components/FilterTextField';
+import AddOrEditStudyModal from 'applications/operationalStudies/components/Study/AddOrEditStudyModal';
 import {
   type PostSearchApiArg,
   osrdEditoastApi,
@@ -19,7 +20,6 @@ import { Loader, Spinner } from 'common/Loaders';
 import SelectionToolbar from 'common/SelectionToolbar';
 import AddOrEditScenarioModal from 'modules/scenario/components/AddOrEditScenarioModal';
 import { cleanScenarioLocalStorage } from 'modules/scenario/helpers/utils';
-import AddOrEditStudyModal from 'modules/study/components/AddOrEditStudyModal';
 import { budgetFormat } from 'utils/numbers';
 
 import DateBox from './components/DateBox';
