+++ front/src/modules/modules.scss
@@ -4,6 +4,5 @@
 @use './rollingStock/styles/rollingStock.scss';
 @use './simulationResult/styles/simulationResult.scss';
 @use './scenario/styles/scenario.scss';
-@use './study/styles/study.scss';
 @use './timetableItem/styles/timetableItem.scss';
 @use './timesStops/styles/timesStops.scss';
