--- front/src/modules/study/styles/study.scss
@@ -1,1 +1,0 @@
-@use './addOrEditStudyModal';
